# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: BeefTV HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /agent/memories:
    get:
      tags: [agent]
      summary: 列出当前用户的 Agent 记忆
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: 信封 data.memories
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [agent]
      summary: 手动添加一条个人记忆（直接批准）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, category, situation]
              properties:
                topic: { type: string }
                category: { type: string }
                situation: { type: string }
                lesson: { type: string }
                source: { type: string }
                steps:
                  type: array
                  items:
                    type: object
                    properties:
                      tool: { type: string }
                      action: { type: string }
                      note: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/export:
    get:
      tags: [agent]
      summary: 导出当前用户的 Agent 记忆
      responses:
        "200":
          description: 信封 data 为 version/kind/exportedAt/memories
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/import:
    post:
      tags: [agent]
      summary: 导入 Agent 记忆 JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: integer }
                kind: { type: string }
                memories: { type: array }
      responses:
        "200":
          description: 信封 data 含 imported、merged、skipped
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/settings:
    get:
      tags: [agent]
      summary: 读取当前用户的记忆压缩设置与最近一次压缩状态
      responses:
        "200":
          description: 信封 data 含 compactInterval、模型字段、lastStatus、summary
        "401":
          $ref: "#/components/responses/Error"
    patch:
      tags: [agent]
      summary: 更新记忆压缩周期和文本模型
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compactInterval:
                  type: string
                  enum: [off, daily, weekly, monthly]
                logicalModelId: { type: string }
                channelId: { type: string }
                channelModelKey: { type: string }
                model: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/compact:
    post:
      tags: [agent]
      summary: 提交一次个人记忆压缩任务（走用户文本模型计费）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logicalModelId: { type: string }
                channelId: { type: string }
                channelModelKey: { type: string }
                model: { type: string }
      responses:
        "200":
          description: 信封 data 为压缩设置视图，lastStatus 为 queued
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [agent]
      summary: 更新一条属于当前用户的记忆
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [agent]
      summary: 删除一条属于当前用户的记忆
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /agent/memories/{id}/decide:
    post:
      tags: [agent]
      summary: 批准或拒绝一条属于当前用户的待审记忆
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /tasks:
    post:
      tags: [tasks]
      summary: 创建生成任务
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /tasks/{id}/text-events:
    get:
      tags: [tasks]
      summary: 文本任务 SSE
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: integer }
          description: 递增事件游标，不要用任务 ID
      responses:
        "200":
          description: text/event-stream
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"